pyproject.toml
[tool.poetry]
name = "link-manager"
version = "0.1.0"
description = ""
authors = ["Siarhey <borodavkasv111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
flake8 = "6.1.0"
pre-commit = "^3.7.0"
python-dotenv = "1.0.1"
black = "^24.4.0"
isort = "^5.13.2"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
drf-yasg = "^1.21.7"
faker = "^24.11.0"
pandas = "^2.2.2"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
